py11; removing unnecessary code referring to BG_DOWNLOAD or KaldiNL HTTP API; restructuring towards simpler direct usage
