+++ pyproject.toml
@@ -3,9 +3,10 @@ name = "dane-asr-worker"
 version = "0.1.0"
 description = "Automatic speech recognition through an external service. Depends on DANE-server"
 authors = ["Jaap Blom <jblom@beeldengeluid.nl>", "Govert Brinkmann <gbrinkmann@beeldengeluid.nl>"]
+package-mode = false
 
 [tool.poetry.dependencies]
-python = "^3.10"
+python = "^3.11"
 requests = "^2.28.1"
 pika = "^1.3.1"
 yacs = "^0.1.8"
@@ -56,7 +57,7 @@ exclude = '''
 '''
 
 [tool.mypy]
-python_version = "3.10"
+python_version = "3.11"
 
 [[tool.mypy.overrides]]
 module = [
